+++ src/qbsprojectmanager.ts
@@ -7,6 +7,7 @@ import * as which from 'which';
 
 import { escapeShell } from './qbsutils';
 import { QbsBuildConfigurationManager } from './qbsbuildconfigurationmanager';
+import { QbsBuildConfigurationName } from './datatypes/qbsbuildconfigurationname';
 import { QbsBuildProfileManager } from './qbsbuildprofilemanager';
 import { QbsBuildSystem, QbsBuildSystemTimeout } from './qbsbuildsystem';
 import { QbsCommandKey } from './datatypes/qbscommandkey';
@@ -371,7 +372,7 @@ export class QbsProjectManager implements vscode.Disposable {
         const projects: any = (storage) ? storage[this.qbsStoredFsProjectsKey] : undefined;
         const project: any = (projects) ? projects[this.project?.getFsPath()] : undefined;
         this.project.setProfileName((project) ? project[this.qbsStoredBuildProfileNameKey] : undefined);
-        this.project.setConfigurationName((project) ? project[this.qbsStoredBuildConfigurationNameKey] : undefined);
+        this.project.setConfigurationName((project) ? project[this.qbsStoredBuildConfigurationNameKey] : QbsBuildConfigurationName.Debug);
         this.project.setBuildProductName((project) ? project[this.qbsStoredBuildProductNameKey] : undefined);
         this.project.setLaunchProductName((project) ? project[this.qbsStoredLaunchProductNameKey] : undefined);
         this.project.setDebuggerName((project) ? project[this.qbsStoredDebuggerNameKey] : undefined);